# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "APIErrorResponse": {
        "description": "The standard COMPASS JSON error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "APIErrorResponse",
        "type": "object"
      },
      "AccountCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "middle_name": {
            "default": "",
            "title": "Middle Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/RoleCode"
          },
          "suffix": {
            "default": "",
            "title": "Suffix",
            "type": "string"
          }
        },
        "required": [
          "email",
          "first_name",
          "last_name",
          "role"
        ],
        "title": "AccountCreateRequest",
        "type": "object"
      },
...
      "AccountListResponse": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "AccountListResponse",
        "type": "object"
      },
...
      "DesignationCode": {
        "enum": [
          "DPO",
          "HEAD_GUIDANCE_COUNSELOR"
        ],
        "title": "DesignationCode",
        "type": "string"
      },
...
      "RoleCode": {
        "enum": [
          "COUNSELOR",
          "GUIDANCE_SERVICES_STAFF",
          "IT_ADMIN",
          "STUDENT"
        ],
        "title": "RoleCode",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OpaqueSessionAuth": {
        "description": "HttpOnly opaque COMPASS session cookie. The browser sends it automatically; it is not a Bearer token. State-changing requests also require Django CSRF protection.",
        "in": "cookie",
        "name": "compass_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The version-one backend API for COMPASS.",
    "title": "COMPASS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "accountsList",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RoleCode"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "designation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DesignationCode"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "List managed accounts",
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "operationId": "accountsCreate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailResponse"
                }
              }
            },
            "description": "Created"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Create a managed account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/accounts/{user_id}": {
      "get": {
        "operationId": "accountsGet",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Inspect a managed account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/accounts/{user_id}/capability-overrides": {
      "get": {
        "operationId": "accountsListCapabilityOverrides",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityOverrideListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "List managed account capability overrides",
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/api/v1/accounts/{user_id}/designations": {
      "get": {
        "operationId": "accountsListDesignations",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignationListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "List managed account designations",
        "tags": [
          "accounts"
        ]
      }
    },
...
  },
  "servers": [],
...
}